+++ plone/dexterity/content.py
@@ -18,7 +18,6 @@ from Products.CMFCore.interfaces import IMutableDublinCore
 from Products.CMFCore.interfaces import ITypeInformation
 from Products.CMFDynamicViewFTI.browserdefault import BrowserDefaultMixin
 from Products.CMFPlone.interfaces import IConstrainTypes
-from copy import deepcopy
 from plone.autoform.interfaces import READ_PERMISSIONS_KEY
 from plone.behavior.interfaces import IBehaviorAssignable
 from plone.dexterity.filerepresentation import DAVCollectionMixin
@@ -28,6 +27,7 @@ from plone.dexterity.interfaces import IDexterityContent
 from plone.dexterity.interfaces import IDexterityItem
 from plone.dexterity.schema import SCHEMA_CACHE
 from plone.dexterity.utils import all_merged_tagged_values_dict
+from plone.dexterity.utils import default_from_schema
 from plone.dexterity.utils import datify
 from plone.dexterity.utils import iterSchemata
 from plone.dexterity.utils import safe_unicode
@@ -44,34 +44,16 @@ from zope.interface.declarations import Implements
 from zope.interface.declarations import ObjectSpecificationDescriptor
 from zope.interface.declarations import getObjectSpecification
 from zope.interface.declarations import implementedBy
-from zope.schema.interfaces import IContextAwareDefaultFactory
 from zope.security.interfaces import IPermission
 
 import six
+import os
 
 
 _marker = object()
 _zone = DateTime().timezone()
 FLOOR_DATE = DateTime(1970, 0)  # always effective
 CEILING_DATE = DateTime(2500, 0)  # never expires
-
-
-def _default_from_schema(context, schema, fieldname):
-    """helper to lookup default value of a field
-    """
-    if schema is None:
-        return _marker
-    field = schema.get(fieldname, None)
-    if field is None:
-        return _marker
-    if IContextAwareDefaultFactory.providedBy(
-            getattr(field, 'defaultFactory', None)
-    ):
-        bound = field.bind(context)
-        return deepcopy(bound.default)
-    else:
-        return deepcopy(field.default)
-    return _marker
 
 
 class FTIAwareSpecification(ObjectSpecificationDescriptor):
@@ -321,9 +303,9 @@ class DexterityContent(DAVResourceMixin, PortalContent, PropertyManager,
         for (k, v) in kwargs.items():
             setattr(self, k, v)
 
-    def __getattr__(self, name):
-        # python basics:  __getattr__ is only invoked if the attribute wasn't
-        # found by __getattribute__
+    def __getattr_default__(self, name):
+        # python basics:  __getattr__ is only invoked if the attribute
+        # wasn't found by __getattribute__
         #
         # optimization: sometimes we're asked for special attributes
         # such as __conform__ that we can disregard (because we
@@ -333,12 +315,13 @@ class DexterityContent(DAVResourceMixin, PortalContent, PropertyManager,
         if name.startswith('__') or name == '_v__providedBy__':
             raise AttributeError(name)
 
-        # attribute was not found; try to look it up in the schema and return
-        # a default
-        value = _default_from_schema(
+        # attribute was not found; try to look it up in the schema and
+        # return a default
+        value = default_from_schema(
             self,
             SCHEMA_CACHE.get(self.portal_type),
-            name
+            name,
+            default=_marker
         )
         if value is not _marker:
             return value
@@ -348,15 +331,19 @@ class DexterityContent(DAVResourceMixin, PortalContent, PropertyManager,
         if assignable is not None:
             for behavior_registration in assignable.enumerateBehaviors():
                 if behavior_registration.interface:
-                    value = _default_from_schema(
+                    value = default_from_schema(
                         self,
                         behavior_registration.interface,
-                        name
+                        name,
+                        default=_marker
                     )
                     if value is not _marker:
                         return value
 
         raise AttributeError(name)
+
+    if not os.environ.get('DEXTERITY_WITHOUT_GETATTR'):
+        __getattr__ = __getattr_default__
 
     # Let __name__ and id be identical. Note that id must be ASCII in Zope 2,
     # but __name__ should be unicode. Note that setting the name to something
@@ -662,8 +649,9 @@ class Item(PasteBehaviourMixin, BrowserDefaultMixin, DexterityContent):
         'action': 'view',
     },) + CMFCatalogAware.manage_options + SimpleItem.manage_options
 
-    # Be explicit about which __getattr__ to use
-    __getattr__ = DexterityContent.__getattr__
+    if not os.environ.get('DEXTERITY_WITHOUT_GETATTR'):
+        # Be explicit about which __getattr__ to use
+        __getattr__ = DexterityContent.__getattr__
 
 
 @implementer(IDexterityContainer)
@@ -703,14 +691,18 @@ class Container(
         CMFOrderedBTreeFolderBase.__init__(self, id)
         DexterityContent.__init__(self, id, **kwargs)
 
-    def __getattr__(self, name):
-        try:
-            return DexterityContent.__getattr__(self, name)
-        except AttributeError:
-            pass
+    if not os.environ.get('DEXTERITY_WITHOUT_GETATTR'):
+        def __getattr__(self, name):
+            try:
+                return DexterityContent.__getattr__(self, name)
+            except AttributeError:
+                pass
 
+            # Be specific about the implementation we use
+            return CMFOrderedBTreeFolderBase.__getattr__(self, name)
+    else:
         # Be specific about the implementation we use
-        return CMFOrderedBTreeFolderBase.__getattr__(self, name)
+        __getattr__ = CMFOrderedBTreeFolderBase.__getattr__
 
     @security.protected(permissions.DeleteObjects)
     def manage_delObjects(self, ids=None, REQUEST=None):