+++ CHANGES.rst
@@ -11,7 +11,10 @@ Breaking changes:
 
 New features:
 
-- *add item here*
+- With environment variable DEXTERITY_WITHOUT_GETATTR being set,
+  Dexterity content no longer has implicit default value lookup
+  on attribute access
+  [datakurre]
 
 Bug fixes: